Add image variable to terraform and yaml files
The commit primarily introduces a new variable for image in the terraform and the yaml files. This image variable corresponds to the image for the container which has a default value added as well in the terraform file. Additionally, redundant quotes around environment variable values have been removed for cleaner code.