Set service type to LoadBalancer and remove ingress resource
The type of the service in the terraform main.tf file has been set to "LoadBalancer". In addition to this, the kubernetes_ingress_v1 resource also known as the 'default', has been removed entirely from the file. This change opts for a LoadBalancer service instead of using ingress, simplifying network management by directly managing external access to the services in a Kubernetes cluster.