Add variable for ingress rule host
This update introduces a new variable, "ingress_rule_host", for configuration of the host in the ingress rule across various applications and environments. This allows for easier modification and more consistent management of the host settings. Changes extend to base, staging, and production versions of both "santa" and "vote" applications.