Comment out ClusterIssuer configuration in cert-manager.yaml
The ClusterIssuer configuration details in the cert-manager.yaml file have been commented out. This change was made to disable the configuration settings for the let's encrypt service. The need to make this change may have been to avoid any potential security issues, or because the service is no longer required.